Organize tests and fixtures, update affected paths
